Dialogs/DownloadFilePicker: Refactor Modal File Download Dialog Refactor the DownloadFile modal download progress dialog out of the DownloadFilePicker.cpp and make universally usable.
